requirements.txt
gmsh 
torch 
pytorch_geometric
meshio
numpy 
matplotlib 
tqdm 
argparse
pandas 
seaborn
toml